DIV-4303 - fixed broken functional tests
Removed autowire for SeparationStrategy as it's created statically
