+++ src/main/java/uk/gov/hmcts/reform/divorce/caseformatterservice/strategy/reasonfordivorce/SeparationTwoYearsStrategy.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.divorce.caseformatterservice.strategy.reasonfordivorce;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.CoreCaseData;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.usersession.DivorceSession;
@@ -13,8 +12,11 @@ public class SeparationTwoYearsStrategy implements ReasonForDivorceStrategy {
 
     private static final String SEPARATION_STRING = "I have been separated from my %s for 2 years or more from the %s.";
 
-    @Autowired
-    public SeparationStrategy separationStrategy;
+    private final SeparationStrategy separationStrategy;
+
+    public SeparationTwoYearsStrategy(SeparationStrategy separationStrategy) {
+        this.separationStrategy = separationStrategy;
+    }
 
     @Override
     public String deriveStatementOfCase(DivorceSession divorceSession) {
