+++ src/test/java/uk/gov/hmcts/reform/divorce/caseformatterservice/strategy/reasonfordivorce/SeparationFiveYearsStrategyUTest.java
@@ -1,6 +1,9 @@
 package uk.gov.hmcts.reform.divorce.caseformatterservice.strategy.reasonfordivorce;
 
+import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
+import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.CoreCaseData;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.usersession.DivorceSession;
 
 import java.text.ParseException;
@@ -9,11 +12,21 @@ import java.util.Locale;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
 
 public class SeparationFiveYearsStrategyUTest {
     private static final String SEPARATION_FIVE_YEARS = "separation-5-years";
 
-    private final SeparationFiveYearsStrategy separationFiveYearsStrategy = new SeparationFiveYearsStrategy();
+    private SeparationStrategy separationStrategy;
+
+    private SeparationFiveYearsStrategy separationFiveYearsStrategy;
+
+    @Before
+    public void setup() {
+        separationStrategy = mock(SeparationStrategy.class);
+        separationFiveYearsStrategy = new SeparationFiveYearsStrategy(separationStrategy);
+    }
 
     @Test
     public void testSeparationFiveYearsStatementOfCase() throws ParseException {
@@ -38,5 +51,34 @@ public class SeparationFiveYearsStrategyUTest {
 
         assertThat(derivedStatementOfCase, equalTo(
             "I have been separated from my null for 5 years or more from the 01 January 1970."));
+    }
+
+    @Test
+    public void testAcceptsMatchesFiiveYearSeparation() {
+        assertTrue(separationFiveYearsStrategy.accepts("sEparAtion-5-Years"));
+    }
+
+    @Test
+    public void testSetLivedApartFieldsFromDivorceSessionCallsSeparationStrategy() {
+        //given
+        DivorceSession divorceSession = new DivorceSession();
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        //when
+        separationFiveYearsStrategy.setLivedApartFieldsFromDivorceSession(divorceSession, coreCaseData);
+
+        Mockito.verify(separationStrategy).setLivedApartFieldsFromDivorceSession(divorceSession, coreCaseData);
+    }
+
+    @Test
+    public void testSetLivedApartFieldsFromCoreCaseDataCallsSeparationStrategy() {
+        //given
+        DivorceSession divorceSession = new DivorceSession();
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        //when
+        separationFiveYearsStrategy.setLivedApartFieldsFromCoreCaseData(coreCaseData, divorceSession);
+
+        Mockito.verify(separationStrategy).setLivedApartFieldsFromCoreCaseData(coreCaseData, divorceSession);
     }
 }
