+++ src/test/java/uk/gov/hmcts/reform/divorce/caseformatterservice/strategy/reasonfordivorce/SeparationTwoYearsStrategyUTest.java
@@ -1,6 +1,9 @@
 package uk.gov.hmcts.reform.divorce.caseformatterservice.strategy.reasonfordivorce;
 
+import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
+import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.CoreCaseData;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.usersession.DivorceSession;
 
 import java.text.ParseException;
@@ -9,11 +12,21 @@ import java.util.Locale;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
 
 public class SeparationTwoYearsStrategyUTest {
     private static final String SEPARATION_TWO_YEARS = "separation-2-years";
 
-    private final SeparationTwoYearsStrategy separationTwoYearsStrategy = new SeparationTwoYearsStrategy();
+    private SeparationStrategy separationStrategy;
+
+    private SeparationTwoYearsStrategy separationTwoYearsStrategy;
+
+    @Before
+    public void setup() {
+        separationStrategy = mock(SeparationStrategy.class);
+        separationTwoYearsStrategy = new SeparationTwoYearsStrategy(separationStrategy);
+    }
 
     @Test
     public void testSeparationTwoYearsStatementOfCase() throws ParseException {
@@ -38,5 +51,34 @@ public class SeparationTwoYearsStrategyUTest {
 
         assertThat(derivedStatementOfCase, equalTo(
             "I have been separated from my null for 2 years or more from the 01 January 1970."));
+    }
+
+    @Test
+    public void testAcceptsMatchesTwoYearSeparation() {
+        assertTrue(separationTwoYearsStrategy.accepts("sEparAtion-2-Years"));
+    }
+
+    @Test
+    public void testSetLivedApartFieldsFromDivorceSessionCallsSeparationStrategy() {
+        //given
+        DivorceSession divorceSession = new DivorceSession();
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        //when
+        separationTwoYearsStrategy.setLivedApartFieldsFromDivorceSession(divorceSession, coreCaseData);
+
+        Mockito.verify(separationStrategy).setLivedApartFieldsFromDivorceSession(divorceSession, coreCaseData);
+    }
+
+    @Test
+    public void testSetLivedApartFieldsFromCoreCaseDataCallsSeparationStrategy() {
+        //given
+        DivorceSession divorceSession = new DivorceSession();
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        //when
+        separationTwoYearsStrategy.setLivedApartFieldsFromCoreCaseData(coreCaseData, divorceSession);
+
+        Mockito.verify(separationStrategy).setLivedApartFieldsFromCoreCaseData(coreCaseData, divorceSession);
     }
 }
